Move Description logic to lib/git/
